(swab_long_long, swab_long): better now.

+++ ufs/ufs.h
@@ -181,14 +181,14 @@ swab_short (short arg)
 extern inline long
 swab_long (long arg)
 {
-  return ((swab_short (arg & 0xffff) << 16)
+  return ((long) (swab_short (arg & 0xffff) << 16)
 	  | swab_short ((arg & 0xffff0000) >> 16));
 }
 
 extern inline long long
 swab_long_long (long long arg)
 {
-  return ((swab_long (arg & 0xffffffff) << 32)
+  return ((long long)(swab_long (arg & 0xffffffff) << 32)
 	  | swab_long ((arg & 0xffffffff00000000) >> 32));
 }
